+++ TDRepo_Engine/Convert/ToBHoM/Node.cs
@@ -1,0 +1,26 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using BH.oM.Structure.Elements;
+
+namespace BH.Engine.TDRepo
+{
+    public static partial class Convert
+    {
+        /***************************************************/
+        /**** Public Methods                            ****/
+        /***************************************************/
+
+        //Add methods for converting to BHoM from the specific software types. 
+        //Only do this if possible to do without any com-calls or similar to the adapter
+        //Example:
+        //public static Node ToBHoM(this _3DRepo_ToolkitNode node)
+        //{
+        //    //Insert code for convertion
+        //}
+
+        /***************************************************/
+    }
+}
