+++ TDRepo_Toolkit.sln
@@ -1,0 +1,37 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.852
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TDRepo_Adapter", "TDRepo_Adapter\TDRepo_Adapter.csproj", "{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TDRepo_Engine", "TDRepo_Engine\TDRepo_Engine.csproj", "{625C4A3B-A0AB-4B29-B289-4C439B565EFF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TDRepo_oM", "TDRepo_oM\TDRepo_oM.csproj", "{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CFD6BA4A-3480-4351-9F7F-B1B37ABB2F21}.Release|Any CPU.Build.0 = Release|Any CPU
+		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{625C4A3B-A0AB-4B29-B289-4C439B565EFF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {25B8320F-DE2C-4CBE-98A5-584D670916FB}
+	EndGlobalSection
+EndGlobal
