+++ TDRepo_oM/TDRepo_oM.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{A70835F4-6498-4F3B-B7BA-5BC87F36AB4C}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>RepoClient</RootNamespace>
-    <AssemblyName>RepoClient</AssemblyName>
+    <RootNamespace>BH.oM.TDRepo</RootNamespace>
+    <AssemblyName>TDRepo_oM</AssemblyName>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
   </PropertyGroup>
@@ -41,10 +41,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Logger.cs" />
+    <Compile Include="Objects\Mesh.cs" />
     <Compile Include="MultipartForm.cs" />
+    <Compile Include="Objects\Point.cs" />
     <Compile Include="RepoController.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RepoObjects.cs" />
+    <Compile Include="Objects\Face.cs" />
     <Compile Include="SceneCreator.cs" />
     <Compile Include="Connector.cs" />
   </ItemGroup>
