--- _3DRepo_Toolkit_Engine/Convert/AdapterId.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BH.Engine._3DRepo_Toolkit
-{
-    public static partial class Convert
-    {
-        //Key used to find the software in the custom data of created or read objects
-        public const string AdapterIdName = "3DRepo_id";
-    }
-}
